AP-13863: remove working dir from FS file system constructor
Except the local file system, all other file system override the
getWorkflowDirectory method and implement it on their own. Its hard to
discover the workflow directory already in the call of the super
constructor. Remove the parameter from the constructor and implement the
getWorkflowDirectory method in the local file system implementation.

AP-13863 ("Relative To" NIO file system rewrite (Server version))